# docs/openapi/paella-openapi-full.yaml
openapi: 3.0.0
info:
  title: 'Billie - PaD API Documentation'
  description: |
    # Introduction
    
    With PayAfterDelivery (PAD) / "Rechnungskauf" Billie offers a convenient, secure and simple way to allow your B2B-customers
    to pay on invoice. Our solution enables merchants to accept invoice payments from other business customers without waiting for their money,
    reminding customers about the payment, or assuming credit risk.
    
    The following documentation describes our RESTFul API, which allows easy integration into your web-shop,
    eCommerce platform or invoicing back-end.
    
    Based upon your Billie contract, your customers will have between 7 and 120 days (payment term) to pay your invoices, while Billie will 
    transfer the money to you immediately after shipment of goods or fulfilment of service obligation. For outstanding payments, 
    Billie takes over reminding customers on your behalf (white-label solution). Additionally, Billie protects you from credit risks 
    by absorbing any potential debtor defaults.
    
    ---
    
    ## Order States
    
    | State         | Description                                                                                   |
    |---------------|-----------------------------------------------------------------------------------------------|
    | waiting       | The order creation failed and needs a manual operation in order to be approved or declined    |
...
    
    ## Order State Transitions
    
...
servers:
...
paths:
...
  '/payments/{uuid}':
    get:
      tags:
        - 'Dashboard Payments'
      summary: 'Get Payment Details'
      operationId: get_payment_details
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: 'Successful response'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMerchantPaymentDetailsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      security:
        -
          oauth2: []
      x-groups:
        - private
...
  /merchant/user:
    get:
      tags:
        - 'Dashboard Users'
      summary: 'Get Merchant User'
      operationId: get_merchant_user
      responses:
        '200':
          description: 'Merchant User details'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMerchantUserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
      security:
        -
          oauth2: []
      x-groups:
        - private
...
  /merchant/user/login:
    post:
      tags:
        - 'Dashboard Users'
      summary: 'Merchant User Login'
      operationId: merchant_user_login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MerchantUserLoginRequest'
      responses:
        '200':
          description: 'Login successful'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantUserLoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
      security: []
      x-groups:
        - private
...
components:
  schemas:
    UUID:
      type: string
      format: uuid
      maxLength: 36
      minLength: 36
      example: 12345667-890a-bcde-f123-34567890abcd
...
    IBAN:
      title: IBAN
      type: string
      maxLength: 22
      minLength: 22
      pattern: '^[A-Z]{2}(?:[ ]?[0-9]){18,20}$'
      example: DE61500105175136458915
    Date:
      title: Date
      type: string
      format: date
      example: '2019-01-20'
...
    TinyText:
      title: string
      type: string
      maxLength: 255
...
    Money:
      title: Money
      type: number
      format: float
      pattern: '^(\d{1,}(.\d{2})?)$'
      example: 99.95
    MerchantUserPermissions:
      title: 'Merchant User Permissions'
      type: string
      enum:
        - VIEW_ORDERS
        - VIEW_DEBTORS
        - VIEW_PAYMENTS
        - CONFIRM_ORDER_PAYMENT
        - PAUSE_DUNNING
      example:
        - VIEW_ORDERS
        - VIEW_DEBTORS
        - VIEW_PAYMENTS
        - CONFIRM_ORDER_PAYMENT
        - PAUSE_DUNNING
...
    GetMerchantUserResponse:
      title: 'Get Merchant User Response'
      properties:
        first_name:
          type: string
          nullable: false
        last_name:
          type: string
          nullable: false
        email:
          type: string
          format: email
          nullable: false
        user_id:
          type: integer
          nullable: false
        merchant_company:
          description: 'Merchant company data.'
          properties:
            name:
              example: 'Billie GmbH'
              nullable: true
              $ref: '#/components/schemas/TinyText'
            address_house_number:
              example: '4'
              nullable: true
              $ref: '#/components/schemas/TinyText'
            address_street:
              example: Charlottenstr.
              nullable: true
              $ref: '#/components/schemas/TinyText'
            address_city:
              example: Berlin
              nullable: true
              $ref: '#/components/schemas/TinyText'
            address_postal_code:
              type: string
              maxLength: 5
              example: '10969'
              nullable: true
            address_country:
              type: string
              maxLength: 2
              nullable: true
          type: object
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/MerchantUserPermissions'
          nullable: false
        role:
          description: 'Internal role name'
          type: string
          example: admin
          nullable: false
      type: object
...
    MerchantUserLoginRequest:
      title: 'Merchant User Login'
      properties:
        email:
          type: string
          format: email
          nullable: false
        password:
          type: string
          format: password
          nullable: false
      type: object
    MerchantUserLoginResponse:
      title: 'Merchant User Login Response'
      properties:
        user_id:
          type: integer
          nullable: false
        access_token:
          type: string
          example: l387435hzyoc0oo4kokow
          nullable: false
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/MerchantUserPermissions'
          nullable: false
        merchant_name:
          type: string
          example: 'Billie GmbH'
          nullable: false
      type: object
...
    ErrorsObject:
      title: 'API Errors'
      properties:
        errors:
          type: array
          items:
            required:
              - title
              - code
            properties:
              title:
                description: 'Error message'
                type: string
              code:
                description: 'Error code'
                type: string
                nullable: true
              source:
                description: 'Property path where the error occurred, if applicable'
                type: string
                nullable: true
            type: object
      type: object
    GetMerchantPaymentDetailsResponse:
      title: 'Merchant Payments Response'
      properties:
        uuid:
          $ref: '#/components/schemas/UUID'
        amount:
          $ref: '#/components/schemas/Money'
        transaction_date:
          $ref: '#/components/schemas/Date'
        is_allocated:
          type: boolean
        overpaid_amount:
          $ref: '#/components/schemas/Money'
        transaction_counterparty_iban:
          $ref: '#/components/schemas/IBAN'
        transaction_counterparty_name:
          type: string
        transaction_reference:
          type: string
        merchant_debtor_uuid:
          $ref: '#/components/schemas/UUID'
        orders:
          type: array
          items:
            properties:
              uuid:
                $ref: '#/components/schemas/UUID'
              amount:
                $ref: '#/components/schemas/Money'
              outstanding_amount:
                $ref: '#/components/schemas/Money'
              external_id:
                type: string
              invoice_number:
                type: string
            type: object
      type: object
...
  responses:
...
    ServerError:
      description: 'Unexpected Server Error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsObject'
    NotFound:
      description: 'Resource Not Found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsObject'
    Unauthorized:
      description: 'Unauthorized request or invalid credentials'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsObject'
    Forbidden:
      description: 'Forbidden Request. The operation cannot be completed.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsObject'
    BadRequest:
      description: 'Invalid Request Data'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorsObject'
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        This authentication mechanism is based on the OAuth 2.0 prototocol and requires a Client ID, a Client Secret Key and
        in some cases also a user and password.
        
        With these credentials, the client should request a token using the `POST /api/v1/oauth/token` endpoint,
        which should be used in the `Authorization` header together with the `Bearer` challenge to authenticate all the requests.
        
        > Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c`.
      flows:
        password:
          tokenUrl: /oauth/token
          scopes:
            default: default
    apiKey:
      type: apiKey
      description: |
        This authentication mechanism is based on the `X-Api-Key` header. It has been is **deprecated**, since it is
        less secure because API Keys never expire, in comparison to OAuth tokens.
        
        > Example: `X-Api-Key: bf428b05-f80e-4ed0-aa79-e0e7bef274a3`.
      name: X-Api-Key
      in: header
      x-groups:
        - internal
...